+++ server/routes/users.js
@@ -941,9 +941,8 @@ router.put(
 
 // User update their password from profile page
 router.put(
-    '/profile/:id/edit-password',
+    '/:studentId/edit-student-password',
     isAuthenticated,
-    editSelfPermission,
     (req, res, next) => {
         // Hash the password.
         bcrypt.hash(
@@ -957,7 +956,7 @@ router.put(
                 // Add data.
                 let sqlQuery = `UPDATE users 
                     SET password = ${conn.escape(hashedPassword)} 
-                    WHERE id = ${conn.escape(req.params.id)};`;
+                    WHERE id = ${conn.escape(req.params.studentId)};`;
 
                 conn.query(sqlQuery, (err, results) => {
                     try {
